gpu-miner: Fixed a severe bug wherein most shares would be rejected as duplicate shares
This bug was caused by the fact that we were multiplexing the same XMRigWork struct but just
changing the NoncePtr values and pushing them off to HashCheckChan to be checked. Since this
happens in a loop, if there were two acceptable hashes within the 0xFF results, we would run
into an issue where multiple results would be submitted with the same hash.

This commit uses the newly created XMRigWork.Clone() function to create a copy of the
XMRigWork struct. A clone is created for each of the 0xFF results and this clone is
then shipped off to HashCheckChan.